다른 사용자가 url주소로 직접 게시물 수정/삭제 시도하지 못하도록 예외처리 추가함

+++ src/main/java/com/fitness/gymsup/Controller/DiaryBoardController.java
@@ -213,9 +213,16 @@ public class DiaryBoardController {
     }
     @GetMapping("/board_diary_modify")
     public String modifyForm(Integer id,
+                             Integer boardUserId,
                              Model model,
                              HttpServletRequest request,
                              Principal principal) throws Exception {
+        log.info("id: "+ id + ", boardUserId:"+boardUserId);
+        if( boardUserId == null ||
+                !boardService.userConfirm(id, request, principal) ) {
+            return "redirect:/";
+        }
+
         BoardDTO boardDTO = boardService.detail(id, false, request, principal);
         model.addAttribute("boardDTO", boardDTO);
         log.info(boardDTO);
@@ -237,8 +244,19 @@ public class DiaryBoardController {
         return "redirect:/board_diary_list";
     }
     @GetMapping("/board_diary_remove")
-    public String removeProc(Integer id, Model model) throws Exception {
+    public String removeProc(Integer id,
+                             Integer boardUserId,
+                             Model model,
+                             HttpServletRequest request,
+                             Principal principal) throws Exception {
+
+        if( boardUserId == null ||
+                !boardService.userConfirm(id, request, principal) ) {
+            return "redirect:/";
+        }
+
         boardService.remove(id);
+
         return "redirect:/board_diary_list";
     }
 }
